--- a/src/api/app/models/owner_search/base.rb
+++ b/src/api/app/models/owner_search/base.rb
@@ -31,7 +31,8 @@ module OwnerSearch
       return [Project.get_by_name(params[:project])] if params[:project]
 
       # Find all marked projects
-      projects = Project.joins(:attribs).where(attribs: { attrib_type_id: attribute.id })
+      projects = nil
+      projects = Project.joins(:attribs).where(attribs: { attrib_type_id: attribute.id }) if attribute.present?
       return projects unless projects.empty?
 
       raise AttributeNotSetError, "The attribute #{attribute.fullname} is not set to define default projects."