@fcrozat 2020-08-12
I'm not 100% sure it is related with this PR but since a few days on IBS, packages which were still returned with bugowner who left the company are no longer returned at all.