@hennevogel 2020-08-14
This breaks `Webui::SearchController`...