@eduardoj 2022-03-04
Introducing the method `for_package` in the model `OwnerSearch::Assignee` will require the task of adapting and adding specs in the currently failing [model tests](https://github.com/openSUSE/open-build-service/blob/master/src/api/spec/models/owner_search_spec.rb).