@adrianschroeter 2025-04-24
@hennevogel can you please take over and look into the webui part?