@adrianschroeter 2025-05-22
k, seems I implemented parts of it multiple times..

Rebase it and dropped some parts, but have not  tested it yet.